Create admin dashboard using HTML & CSS
This commit completes the admin dashboard project.

The layout was done completly using grid. Other notable features are
the semi reactive design and the SVGs that were added.

The main issue present is that the header and sidebar are out of order
in the DOM and are repositioned using CSS grid. I don't believe that
this will be an issue for screen readers, however, it might be. This
was done so that adding the box-shadow to the headerwas displayed on top
of the main body. I tried shifting around the CSS to have this same
affect but it didn't work.

Another error is that the images are barely dynamically styles so they
might look wierd in larger screen sizes.